+++ common/utils/src/lib.rs
@@ -12,6 +12,7 @@ pub mod byte_utils;
 
 pub mod byte_nums;
 pub mod duration;
+pub mod id_generator;
 pub mod precision;
 
 pub type Timestamp = i64;
