+++ tskv/src/compaction/job.rs
@@ -3,19 +3,18 @@ use std::sync::atomic::{AtomicBool, AtomicUsize};
 use std::sync::{atomic, Arc};
 use std::time::{Duration, Instant};
 
-use metrics::metric_register::MetricsRegister;
 use snafu::ResultExt;
-use tokio::runtime::Runtime;
 use tokio::sync::mpsc::Receiver;
 use tokio::sync::oneshot::Receiver as OneshotReceiver;
 use tokio::sync::{oneshot, Mutex, Notify, RwLock, RwLockWriteGuard, Semaphore};
-use trace::{error, info, warn};
+use trace::{error, info};
 
 use crate::compaction::metrics::{CompactionType, VnodeCompactionMetrics};
 use crate::compaction::{flush, pick_compaction, CompactTask, FlushReq};
 use crate::error::{CommonSnafu, IndexErrSnafu};
 use crate::mem_cache::memcache::MemCache;
-use crate::summary::SummaryTask;
+use crate::tsfamily::summary::{Summary, SummaryRequest, SummaryTask};
+use crate::version_set::VersionSet;
 use crate::{TsKvContext, TskvResult, VersionEdit, VnodeId};
 
 const COMPACT_BATCH_CHECKING_SECONDS: u64 = 1;
@@ -79,28 +78,18 @@ pub struct CompactJob {
 }
 
 impl CompactJob {
-    pub fn new(
-        runtime: Arc<Runtime>,
-        ctx: Arc<TsKvContext>,
-        metrics_register: Arc<MetricsRegister>,
-    ) -> Self {
+    pub fn new(ctx: Arc<TsKvContext>, version_set: Arc<RwLock<VersionSet>>) -> Self {
         Self {
-            inner: Arc::new(RwLock::new(CompactJobInner::new(
-                runtime,
-                ctx,
-                metrics_register,
-            ))),
+            inner: Arc::new(RwLock::new(CompactJobInner::new(ctx, version_set))),
         }
     }
 
-    pub async fn start_merge_compact_task_job(&self, compact_task_receiver: Receiver<CompactTask>) {
+    pub async fn start_jobs(&self, compact_task_receiver: Receiver<CompactTask>) {
         self.inner
             .read()
             .await
             .start_merge_compact_task_job(compact_task_receiver);
-    }
 
-    pub async fn start_vnode_compaction_job(&self) {
         self.inner.read().await.start_vnode_compaction_job();
     }
 }
@@ -113,25 +102,19 @@ impl std::fmt::Debug for CompactJob {
 
 struct CompactJobInner {
     ctx: Arc<TsKvContext>,
-    runtime: Arc<Runtime>,
-    metrics_register: Arc<MetricsRegister>,
+    version_set: Arc<RwLock<VersionSet>>,
     compact_processor: Arc<RwLock<CompactProcessor>>,
     enable_compaction: Arc<AtomicBool>,
     running_compactions: Arc<AtomicUsize>,
 }
 
 impl CompactJobInner {
-    fn new(
-        runtime: Arc<Runtime>,
-        ctx: Arc<TsKvContext>,
-        metrics_register: Arc<MetricsRegister>,
-    ) -> Self {
+    fn new(ctx: Arc<TsKvContext>, version_set: Arc<RwLock<VersionSet>>) -> Self {
         let compact_processor = Arc::new(RwLock::new(CompactProcessor::default()));
 
         Self {
             ctx,
-            runtime,
-            metrics_register,
+            version_set,
             compact_processor,
             enable_compaction: Arc::new(AtomicBool::new(false)),
             running_compactions: Arc::new(AtomicUsize::new(0)),
@@ -141,7 +124,7 @@ impl CompactJobInner {
     fn start_merge_compact_task_job(&self, mut compact_task_receiver: Receiver<CompactTask>) {
         info!("Compaction: start merge compact task job");
         let compact_processor = self.compact_processor.clone();
-        self.runtime.spawn(async move {
+        self.ctx.runtime.spawn(async move {
             while let Some(compact_task) = compact_task_receiver.recv().await {
                 compact_processor.write().await.insert(compact_task);
             }
@@ -164,17 +147,17 @@ impl CompactJobInner {
             return;
         }
 
-        let runtime_inner = self.runtime.clone();
         let compact_processor = self.compact_processor.clone();
         let enable_compaction = self.enable_compaction.clone();
         let running_compaction = self.running_compactions.clone();
-        let ctx = self.ctx.clone();
-        let metrics_registry = self.metrics_register.clone();
+        let metrics_registry = self.ctx.metrics.clone();
 
-        self.runtime.spawn(async move {
+        let ctx1 = self.ctx.clone();
+        let version_set = self.version_set.clone();
+        self.ctx.runtime.spawn(async move {
             // TODO: Concurrent compactions should not over argument $cpu.
             let compaction_limit = Arc::new(Semaphore::new(
-                ctx.options.storage.max_concurrent_compaction as usize,
+                ctx1.options.storage.max_concurrent_compaction as usize,
             ));
             let mut check_interval =
                 tokio::time::interval(Duration::from_secs(COMPACT_BATCH_CHECKING_SECONDS));
@@ -199,29 +182,26 @@ impl CompactJobInner {
                 info!("Compacting on vnode(job start): {:?}", &vnode_ids_for_debug);
                 for (task, limit) in vnode_ids {
                     let vnode_id = task.vnode_id();
-                    let ts_family = ctx
-                        .version_set
-                        .read()
-                        .await
-                        .get_tsfamily_by_tf_id(vnode_id)
-                        .await;
-                    if let Some(tsf) = ts_family {
+                    let vnode = version_set.read().await.get_vnode(vnode_id).cloned();
+                    if let Some(vnode) = vnode {
                         info!("Starting compaction on ts_family {}", vnode_id);
+                        let tsf = vnode.get_ts_family();
                         if !tsf.read().await.can_compaction() {
                             info!("forbidden compaction on moving vnode {}", vnode_id);
                             return;
                         }
                         let version = tsf.read().await.version();
                         let compact_req = pick_compaction(task, version).await;
-                        if let Some(req) = compact_req {
+                        if let Some(mut req) = compact_req {
+                            req.set_file_id(vnode.get_summary().read().await.file_id());
                             // Method acquire_owned() will return AcquireError if the semaphore has been closed.
                             let permit = compaction_limit.clone().acquire_owned().await.unwrap();
                             let enable_compaction = enable_compaction.clone();
                             let running_compaction = running_compaction.clone();
-
-                            let ctx = ctx.clone();
                             let metrics_registry = metrics_registry.clone();
-                            runtime_inner.spawn(async move {
+
+                            let ctx2 = ctx1.clone();
+                            ctx1.runtime.spawn(async move {
                                 let _guard = limit.lock().await;
                                 // Check enable compaction
                                 if !enable_compaction.load(atomic::Ordering::SeqCst) {
@@ -235,32 +215,27 @@ impl CompactJobInner {
 
                                 let vnode_compaction_metrics = VnodeCompactionMetrics::new(
                                     &metrics_registry,
-                                    ctx.options.storage.node_id,
+                                    ctx2.options.storage.node_id,
                                     vnode_id,
                                     CompactionType::Normal,
-                                    ctx.options.storage.collect_compaction_metrics,
+                                    ctx2.options.storage.collect_compaction_metrics,
                                 );
-                                match super::run_compaction_job(
-                                    req,
-                                    ctx.global_ctx.clone(),
-                                    vnode_compaction_metrics,
-                                )
-                                .await
+                                match super::run_compaction_job(req, vnode_compaction_metrics).await
                                 {
                                     Ok(Some((version_edit, file_metas))) => {
-                                        let (summary_tx, _summary_rx) = oneshot::channel();
-                                        let _ = ctx
-                                            .summary_task_sender
-                                            .send(SummaryTask::new(
-                                                tsf.clone(),
-                                                version_edit,
-                                                Some(file_metas),
-                                                None,
-                                                summary_tx,
-                                            ))
+                                        let request = SummaryRequest {
+                                            version_edit,
+                                            mem_caches: None,
+                                            file_metas: Some(file_metas),
+                                            ts_family: tsf,
+                                        };
+                                        let summary = vnode.get_summary();
+                                        let result = summary
+                                            .write()
+                                            .await
+                                            .apply_version_edit(&request)
                                             .await;
-
-                                        // TODO Handle summary result using summary_rx.
+                                        info!("vnode compaction apply version edit: {:?}", result);
                                     }
                                     Ok(None) => {
                                         info!("There is nothing to compact.");
@@ -309,15 +284,17 @@ impl<F: FnOnce()> Drop for DeferGuard<F> {
 
 pub struct FlushJob {
     ctx: Arc<TsKvContext>,
+    summary: Arc<RwLock<Summary>>,
 
     notify: Arc<Notify>,
     queue: Arc<RwLock<BTreeMap<u64, (FlushReq, SummaryTask)>>>,
 }
 
 impl FlushJob {
-    pub fn new(ctx: Arc<TsKvContext>) -> Arc<Self> {
+    pub fn new(ctx: Arc<TsKvContext>, summary: Arc<RwLock<Summary>>) -> Arc<Self> {
         let job = Self {
             ctx,
+            summary,
             notify: Arc::new(Notify::new()),
             queue: Arc::new(RwLock::new(BTreeMap::new())),
         };
@@ -339,12 +316,13 @@ impl FlushJob {
                 }
 
                 info!("Flush: completion request {} at {}", flush_req, key);
-                if let Err(e) = job.ctx.summary_task_sender.send(summary_task).await {
-                    warn!(
-                        "Flush: failed to send summary task for tsf_id: {}: {e}",
-                        flush_req.tf_id
-                    );
-                }
+                let result = job
+                    .summary
+                    .write()
+                    .await
+                    .apply_version_edit(&summary_task.request)
+                    .await;
+                info!("Flush: request {} apply version {:?}", flush_req, result);
 
                 if flush_req.trigger_compact {
                     let _ = job
