+++ tskv/src/kvcore.rs
@@ -1,9 +1,8 @@
-use std::collections::HashMap;
 use std::sync::Arc;
 use std::time::Duration;
 
 use datafusion::arrow::record_batch::RecordBatch;
-use memory_pool::{MemoryPool, MemoryPoolRef};
+use memory_pool::MemoryPoolRef;
 use meta::error::MetaError;
 use meta::model::MetaRef;
 use metrics::metric_register::MetricsRegister;
@@ -14,8 +13,8 @@ use models::{SeriesId, SeriesKey};
 use snafu::ResultExt;
 use tokio::runtime::Runtime;
 use tokio::sync::broadcast::{self, Sender as BroadcastSender};
-use tokio::sync::mpsc::{self, Receiver, Sender};
-use tokio::sync::{oneshot, RwLock};
+use tokio::sync::mpsc::{self, Sender};
+use tokio::sync::RwLock;
 use trace::{debug, error, info, warn};
 
 use crate::compaction::job::CompactJob;
@@ -27,25 +26,15 @@ use crate::file_system::async_filesystem::LocalFileSystem;
 use crate::file_system::FileSystem;
 use crate::index::IndexResult;
 use crate::kv_option::{Options, StorageOptions};
-use crate::summary::{Summary, SummaryTask};
+use crate::tsfamily::summary::{Summary, SummaryRequest};
 use crate::tsfamily::super_version::SuperVersion;
-use crate::tsfamily::tseries_family::TseriesFamily;
-use crate::version_set::VersionSet;
+use crate::version_set::{split_to_tsfamily, VersionSet};
 use crate::vnode_store::VnodeStorage;
 use crate::{file_utils, Engine, TsKvContext};
 
-// TODO: A small summay channel capacity can cause a block
-pub const COMPACT_REQ_CHANNEL_CAP: usize = 1024;
-pub const SUMMARY_REQ_CHANNEL_CAP: usize = 1024;
-
 pub struct TsKv {
     ctx: Arc<TsKvContext>,
-    meta_manager: MetaRef,
-    compact_job: CompactJob,
-    runtime: Arc<Runtime>,
-    vnodes: Arc<RwLock<HashMap<VnodeId, VnodeStorage>>>,
-    metrics: Arc<MetricsRegister>,
-    _memory_pool: Arc<dyn MemoryPool>,
+    version_set: Arc<RwLock<VersionSet>>,
     close_sender: BroadcastSender<Sender<()>>,
 }
 
@@ -57,151 +46,76 @@ impl TsKv {
         memory_pool: MemoryPoolRef,
         metrics: Arc<MetricsRegister>,
     ) -> TskvResult<TsKv> {
-        let (compact_task_sender, compact_task_receiver) = mpsc::channel(COMPACT_REQ_CHANNEL_CAP);
-        let (summary_task_sender, summary_task_receiver) = mpsc::channel(SUMMARY_REQ_CHANNEL_CAP);
-
-        let shared_options = Arc::new(options);
-        let (version_set, summary) = Self::recover_summary(
-            runtime.clone(),
-            memory_pool.clone(),
-            meta_manager.clone(),
-            shared_options.clone(),
-            metrics.clone(),
-        )
-        .await;
-
+        let options = Arc::new(options);
+        let (compact_task_sender, compact_task_receiver) = mpsc::channel(1024);
         let ctx = Arc::new(TsKvContext {
-            version_set,
+            metrics,
+            memory_pool,
+            meta: meta_manager,
             compact_task_sender,
-            summary_task_sender,
+            options: options.clone(),
             runtime: runtime.clone(),
-            options: shared_options.clone(),
-            global_ctx: summary.global_context(),
         });
 
+        let old_summary = file_utils::make_summary_file(options.storage.summary_dir(), 0);
+        if LocalFileSystem::try_exists(&old_summary) {
+            split_to_tsfamily(ctx.clone(), old_summary).await.unwrap();
+            std::fs::remove_dir_all(options.storage.summary_dir()).unwrap();
+        }
+
+        let version_set = Arc::new(RwLock::new(VersionSet::new(ctx.clone())));
+
+        let compact_job = CompactJob::new(ctx.clone(), version_set.clone());
+        compact_job.start_jobs(compact_task_receiver).await;
+        Self::run_flush_cold_vnode_job(ctx.clone(), version_set.clone());
+
         let (close_sender, _close_receiver) = broadcast::channel(1);
-        let compact_job = CompactJob::new(runtime.clone(), ctx.clone(), metrics.clone());
         let core = Self {
             ctx,
-            meta_manager,
-            _memory_pool: memory_pool,
-            compact_job,
+            version_set,
             close_sender,
-            metrics,
-            runtime,
-            vnodes: Default::default(),
         };
 
-        core.run_summary_job(summary, summary_task_receiver);
-        core.run_flush_cold_vnode_job();
-        core.compact_job
-            .start_merge_compact_task_job(compact_task_receiver)
-            .await;
-        core.compact_job.start_vnode_compaction_job().await;
         Ok(core)
     }
 
-    pub fn context(&self) -> Arc<TsKvContext> {
-        self.ctx.clone()
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn recover_summary(
-        runtime: Arc<Runtime>,
-        memory_pool: MemoryPoolRef,
-        meta: MetaRef,
-        opt: Arc<Options>,
-        metrics: Arc<MetricsRegister>,
-    ) -> (Arc<RwLock<VersionSet>>, Summary) {
-        let summary_dir = opt.storage.summary_dir();
-        LocalFileSystem::create_dir_if_not_exists(Some(&summary_dir)).unwrap();
-
-        let summary_file = file_utils::make_summary_file(&summary_dir, 0);
-        let summary = if LocalFileSystem::try_exists(&summary_file) {
-            Summary::recover(meta, opt, runtime, memory_pool, metrics.clone())
-                .await
-                .unwrap()
-        } else {
-            Summary::new(opt, runtime, meta, memory_pool, metrics)
-                .await
-                .unwrap()
-        };
-        let version_set = summary.version_set();
-
-        (version_set, summary)
-    }
-
-    fn run_summary_job(
-        &self,
-        mut summary: Summary,
-        mut summary_task_receiver: Receiver<SummaryTask>,
-    ) {
-        self.runtime.spawn(async move {
-            let mut roll_file_time_stamp = models::utils::now_timestamp_secs();
-
-            while let Some(task) = summary_task_receiver.recv().await {
-                info!("Apply Summary task: {:?}", task.request.version_edit);
-                let result = summary.apply_version_edit(&task.request).await;
-                if let Err(err) = task.call_back.send(result) {
-                    trace::info!("Response apply version edit failed: {:?}", err);
-                }
-
-                // Try to rolling summary file every 10mins
-                if models::utils::now_timestamp_secs() - roll_file_time_stamp > 10 * 60 {
-                    if let Err(err) = summary.roll_summary_file().await {
-                        error!("roll summary file failed: {:?}", err);
-                    } else {
-                        roll_file_time_stamp = models::utils::now_timestamp_secs();
-                    }
-                }
-            }
-        });
-
-        info!("Summary task handler started");
-    }
-
-    fn run_flush_cold_vnode_job(&self) {
-        let tskv_ctx = self.ctx.clone();
-        let compact_trigger_cold_duration = tskv_ctx.options.storage.compact_trigger_cold_duration;
-        let compact_trigger_file_num = tskv_ctx.options.storage.compact_trigger_file_num;
-        let compact_task_sender = tskv_ctx.compact_task_sender.clone();
+    fn run_flush_cold_vnode_job(ctx: Arc<TsKvContext>, version_set: Arc<RwLock<VersionSet>>) {
+        let compact_trigger_cold_duration = ctx.options.storage.compact_trigger_cold_duration;
+        let compact_trigger_file_num = ctx.options.storage.compact_trigger_file_num;
+        let compact_task_sender = ctx.compact_task_sender.clone();
         if compact_trigger_cold_duration == Duration::ZERO {
             return;
         }
 
-        let vnodes = self.vnodes.clone();
-        self.runtime.spawn(async move {
+        ctx.runtime.spawn(async move {
             let mut cold_check_interval = tokio::time::interval(Duration::from_secs(60));
             loop {
                 cold_check_interval.tick().await;
 
-                let dbs = tskv_ctx.version_set.read().await.get_all_db().clone();
-                for (_, db) in dbs {
-                    let ts_families = db.read().await.ts_families().clone();
-                    for (tf_id, ts_family) in ts_families {
-                        let last_modified = ts_family.read().await.get_last_modified().await;
-                        if let Some(last_modified) = last_modified {
-                            if last_modified.elapsed() < compact_trigger_cold_duration {
-                                continue;
-                            }
+                let vnodes = version_set.read().await.vnodes();
+                for (tf_id, vnode) in vnodes.iter() {
+                    let ts_family = vnode.ts_family();
+                    let last_modified = ts_family.read().await.get_last_modified().await;
+                    if let Some(last_modified) = last_modified {
+                        if last_modified.elapsed() < compact_trigger_cold_duration {
+                            continue;
+                        }
 
-                            let vnode_opt = vnodes.read().await.get(&tf_id).cloned();
-                            if let Some(vnode) = vnode_opt {
-                                if let Err(e) = vnode.flush(true, true, true).await {
-                                    trace::error!("flush code vnode {} faild: {:?}", tf_id, e)
-                                }
-                            }
+                        if let Err(e) = vnode.flush(true, true, true).await {
+                            trace::error!("flush code vnode {} faild: {:?}", tf_id, e);
                         }
-                        let version = ts_family.read().await.super_version().version.clone();
-                        let levels = version.levels_info();
-                        let task = if levels[0].files.len() as u32 >= compact_trigger_file_num {
-                            CompactTask::Delta(tf_id)
-                        } else {
-                            CompactTask::Normal(tf_id)
-                        };
-                        if let Err(e) = compact_task_sender.send(task).await {
-                            warn!("Scheduler(vnode: {tf_id}): Failed to send compact task: {task}: {e}");
-                        }
+                    }
+
+                    let version = ts_family.read().await.super_version().version.clone();
+                    let levels = version.levels_info();
+                    let task = if levels[0].files.len() as u32 >= compact_trigger_file_num {
+                        CompactTask::Delta(*tf_id)
+                    } else {
+                        CompactTask::Normal(*tf_id)
+                    };
+
+                    if let Err(e) = compact_task_sender.send(task).await {
+                        warn!("Failed to send compact task: {task}: {e}");
                     }
                 }
             }
@@ -209,15 +123,19 @@ impl TsKv {
     }
 
     async fn sync_indexs(&self) -> IndexResult<()> {
-        let vnodes_guard = self.vnodes.read().await;
-        for (_, vnode_storage) in vnodes_guard.iter() {
+        let vs_guard = self.version_set.read().await;
+        for (_, vnode_storage) in vs_guard.vnodes().iter() {
             vnode_storage.sync_index().await;
         }
         Ok(())
     }
 
     pub async fn get_db(&self, tenant: &str, database: &str) -> Option<Arc<RwLock<Database>>> {
-        self.ctx.version_set.read().await.get_db(tenant, database)
+        self.version_set.read().await.get_db(tenant, database)
+    }
+
+    pub fn get_ctx(&self) -> Arc<TsKvContext> {
+        self.ctx.clone()
     }
 
     pub(crate) async fn get_db_or_else_create(
@@ -225,12 +143,13 @@ impl TsKv {
         tenant: &str,
         db_name: &str,
     ) -> TskvResult<Arc<RwLock<Database>>> {
-        if let Some(db) = self.ctx.version_set.read().await.get_db(tenant, db_name) {
+        if let Some(db) = self.version_set.read().await.get_db(tenant, db_name) {
             return Ok(db);
         }
 
         let db_schema = self
-            .meta_manager
+            .ctx
+            .meta
             .tenant_meta(tenant)
             .await
             .ok_or(MetaError::TenantNotFound {
@@ -244,27 +163,8 @@ impl TsKv {
             })
             .context(MetaSnafu)?;
 
-        let db = self
-            .ctx
-            .version_set
-            .write()
-            .await
-            .create_db(db_schema)
-            .await?;
+        let db = self.version_set.write().await.create_db(db_schema).await?;
         Ok(db)
-    }
-
-    pub(crate) async fn get_tsfamily_or_else_create(
-        &self,
-        id: VnodeId,
-        db: Arc<RwLock<Database>>,
-    ) -> TskvResult<Arc<RwLock<TseriesFamily>>> {
-        let mut db = db.write().await;
-        if let Some(tf) = db.get_tsfamily(id) {
-            return Ok(tf);
-        }
-
-        db.create_tsfamily(id, self.ctx.clone()).await
     }
 }
 
@@ -276,15 +176,43 @@ impl Engine for TsKv {
         db_name: &str,
         vnode_id: VnodeId,
     ) -> TskvResult<VnodeStorage> {
+        if let Some(vnode) = self.version_set.read().await.get_vnode(vnode_id) {
+            return Ok(vnode.clone());
+        }
+
         let database = self.get_db_or_else_create(tenant, db_name).await?;
 
-        let ts_index = database.write().await.get_ts_index_or_add(vnode_id).await?;
-        let ts_family = self
-            .get_tsfamily_or_else_create(vnode_id, database.clone())
+        let mut vs_lock = self.version_set.write().await;
+        if let Some(vnode) = vs_lock.get_vnode(vnode_id) {
+            return Ok(vnode.clone());
+        }
+
+        let owner = make_owner(tenant, db_name);
+        let dir = self.ctx.options.storage.ts_family_dir(&owner, vnode_id);
+        let summary_file = file_utils::make_tsfamily_summary_file(dir);
+        let summary = Summary::new(vnode_id, self.ctx.clone(), summary_file).await?;
+        let summary = Arc::new(RwLock::new(summary));
+        let ts_family = database
+            .write()
+            .await
+            .get_tsfamily_or_create(vnode_id, summary.clone())
+            .await?;
+        let ts_index = database
+            .write()
+            .await
+            .get_tsindex_or_create(vnode_id)
             .await?;
 
-        let vnode = VnodeStorage::new(vnode_id, database, ts_index, ts_family, self.ctx.clone());
-        self.vnodes.write().await.insert(vnode_id, vnode.clone());
+        let vnode = VnodeStorage::new(
+            self.ctx.clone(),
+            vnode_id,
+            database,
+            summary,
+            ts_index,
+            ts_family,
+        );
+
+        vs_lock.add_vnode(vnode_id, vnode.clone());
 
         Ok(vnode)
     }
@@ -295,32 +223,9 @@ impl Engine for TsKv {
         database: &str,
         vnode_id: VnodeId,
     ) -> TskvResult<()> {
-        self.vnodes.write().await.remove(&vnode_id);
-
-        if let Some(db) = self.ctx.version_set.read().await.get_db(tenant, database) {
-            let mut db_wlock = db.write().await;
-            db_wlock.del_ts_index(vnode_id);
-            db_wlock
-                .del_tsfamily(vnode_id, self.ctx.summary_task_sender.clone())
-                .await;
-
-            let ts_dir = self
-                .ctx
-                .options
-                .storage
-                .ts_family_dir(&make_owner(tenant, database), vnode_id);
-            match std::fs::remove_dir_all(&ts_dir) {
-                Ok(()) => {
-                    info!("Removed TsFamily directory '{}'", ts_dir.display());
-                }
-                Err(e) => {
-                    error!(
-                        "Failed to remove TsFamily directory '{}': {}",
-                        ts_dir.display(),
-                        e
-                    );
-                }
-            }
+        self.version_set.write().await.remove_vnode(vnode_id);
+        if let Some(db) = self.version_set.read().await.get_db(tenant, database) {
+            db.write().await.del_tsfamily_index(vnode_id);
         }
 
         Ok(())
@@ -333,7 +238,7 @@ impl Engine for TsKv {
         vnode_id: VnodeId,
         trigger_compact: bool,
     ) -> TskvResult<()> {
-        let vnode_opt = self.vnodes.read().await.get(&vnode_id).cloned();
+        let vnode_opt = self.version_set.read().await.get_vnode(vnode_id).cloned();
         if let Some(vnode) = vnode_opt {
             vnode.flush(true, true, trigger_compact).await?;
         }
@@ -349,7 +254,7 @@ impl Engine for TsKv {
         vnode_id: VnodeId,
         filter: &ColumnDomains<String>,
     ) -> TskvResult<Vec<SeriesId>> {
-        let (schema, ts_index) = match self.ctx.version_set.read().await.get_db(tenant, database) {
+        let (schema, ts_index) = match self.version_set.read().await.get_db(tenant, database) {
             Some(db) => {
                 let db = db.read().await;
                 let schema = match db.get_table_schema(tab).await? {
@@ -382,7 +287,7 @@ impl Engine for TsKv {
         vnode_id: VnodeId,
         series_id: &[SeriesId],
     ) -> TskvResult<Vec<SeriesKey>> {
-        if let Some(db) = self.ctx.version_set.read().await.get_db(tenant, database) {
+        if let Some(db) = self.version_set.read().await.get_db(tenant, database) {
             Ok(db
                 .read()
                 .await
@@ -400,23 +305,12 @@ impl Engine for TsKv {
         database: &str,
         vnode_id: VnodeId,
     ) -> TskvResult<Option<Arc<SuperVersion>>> {
-        let version_set = self.ctx.version_set.read().await;
-        // Comment it, It's not a error, Maybe the data not right!
-        // if !version_set.db_exists(tenant, database) {
-        //     return Err(SchemaError::DatabaseNotFound {
-        //         database: database.to_string(),
-        //     }
-        //     .into());
-        // }
-        if let Some(tsf) = version_set
-            .get_tsfamily_by_name_id(tenant, database, vnode_id)
-            .await
-        {
-            Ok(Some(tsf.read().await.super_version()))
+        if let Some(vnode) = self.version_set.read().await.get_vnode(vnode_id) {
+            Ok(Some(vnode.ts_family().read().await.super_version()))
         } else {
             debug!(
-                "ts_family {} with db name '{}' not found.",
-                vnode_id, database
+                "not found ts_family {} in {}.{}",
+                vnode_id, tenant, database
             );
             Ok(None)
         }
@@ -428,14 +322,9 @@ impl Engine for TsKv {
 
     async fn compact(&self, vnode_ids: Vec<VnodeId>) -> TskvResult<()> {
         for vnode_id in vnode_ids {
-            if let Some(ts_family) = self
-                .ctx
-                .version_set
-                .read()
-                .await
-                .get_tsfamily_by_tf_id(vnode_id)
-                .await
-            {
+            if let Some(vnode) = self.version_set.read().await.get_vnode(vnode_id) {
+                let ts_family = vnode.ts_family().clone();
+
                 // TODO: stop current and prevent next flush and compaction.
                 if !ts_family.read().await.can_compaction() {
                     warn!("forbidden compaction on moving vnode {}", vnode_id);
@@ -450,36 +339,26 @@ impl Engine for TsKv {
                 }
 
                 let version = ts_family.read().await.version();
-                if let Some(req) = pick_compaction(CompactTask::Manual(vnode_id), version).await {
+                if let Some(mut req) = pick_compaction(CompactTask::Manual(vnode_id), version).await
+                {
+                    req.set_file_id(vnode.get_summary().read().await.file_id());
                     let vnode_compaction_metrics = VnodeCompactionMetrics::new(
-                        &self.metrics,
+                        &self.ctx.metrics,
                         self.ctx.options.storage.node_id,
                         vnode_id,
                         CompactionType::Manual,
                         self.ctx.options.storage.collect_compaction_metrics,
                     );
-                    match compaction::run_compaction_job(
-                        req,
-                        self.ctx.global_ctx.clone(),
-                        vnode_compaction_metrics,
-                    )
-                    .await
-                    {
+                    match compaction::run_compaction_job(req, vnode_compaction_metrics).await {
                         Ok(Some((version_edit, file_metas))) => {
-                            let (summary_tx, _summary_rx) = oneshot::channel();
-                            let _ = self
-                                .ctx
-                                .summary_task_sender
-                                .send(SummaryTask::new(
-                                    ts_family.clone(),
-                                    version_edit,
-                                    Some(file_metas),
-                                    None,
-                                    summary_tx,
-                                ))
-                                .await;
-
-                            // let _ = summary_rx.await;
+                            let request = SummaryRequest {
+                                version_edit,
+                                mem_caches: None,
+                                file_metas: Some(file_metas),
+                                ts_family: ts_family.clone(),
+                            };
+                            let summary = vnode.get_summary();
+                            summary.write().await.apply_version_edit(&request).await?;
                         }
                         Ok(None) => {
                             info!("There is nothing to compact.");
@@ -496,18 +375,15 @@ impl Engine for TsKv {
     }
 
     async fn get_vnode_hash_tree(&self, vnode_id: VnodeId) -> TskvResult<RecordBatch> {
-        for database in self.ctx.version_set.read().await.get_all_db().values() {
-            let db = database.read().await;
-            if let Some(ts_family) = db.ts_families().get(&vnode_id).cloned() {
-                drop(db);
+        let vnode_opt = self.version_set.read().await.get_vnode(vnode_id).cloned();
+        if let Some(vnode) = vnode_opt {
+            let ts_family = vnode.get_ts_family();
+            let owner = ts_family.read().await.owner();
+            let (tenant, db_name) = split_owner(&owner);
+            self.flush_tsfamily(tenant, db_name, vnode_id, false)
+                .await?;
 
-                let owner = ts_family.read().await.owner();
-                let (tenant, db_name) = split_owner(&owner);
-                self.flush_tsfamily(tenant, db_name, vnode_id, false)
-                    .await?;
-
-                return check::vnode_checksum(ts_family).await;
-            }
+            return check::vnode_checksum(ts_family).await;
         }
 
         Ok(RecordBatch::new_empty(check::vnode_table_checksum_schema()))
