+++ tskv/src/tsfamily/mod.rs
@@ -1,6 +1,7 @@
 pub mod cache_group;
 pub mod column_file;
 pub mod level_info;
+pub mod summary;
 pub mod super_version;
 pub mod tseries_family;
 pub mod tsf_metrics;
